+++ hT4_3/src/Const.java
@@ -2,13 +2,13 @@
  * Created by delf on 18.03.14.
  */
 public class Const extends Polynom{
-    protected final double Value;
+    protected final int Value;
 
-    Const(double constValue){
+    Const(int constValue){
         Value=constValue;
     }
 
-    public double evaluate(double var1, double var2, double var3) {
+    public int evaluate(int var1, int var2, int var3) {
         return Value;
     }
 }
