+++ hT4_3/src/Subtract.java
@@ -5,14 +5,14 @@ public class Subtract extends Operation {
     public static final int MIN_OPERANDS_COUNT = 2;
 
     Subtract(Expression3... items) {
-        super(items);
+       super(items);
     }
 
 
 
-    protected double consider(double... items) {
+    protected int consider(int... items) {
         assert items.length >= MIN_OPERANDS_COUNT;
-        double ret = items[0];
+        int ret = items[0];
         for (int i = 1; i < items.length; i++) {
             ret -= items[i];
         }
